Add Check before deleting a profile photo (#959)
* Add Check before deleting a profile photo

* Formatting

* Update HasProfilePhoto.php

+++ src/HasProfilePhoto.php
@@ -41,6 +41,10 @@ trait HasProfilePhoto
             return;
         }
 
+        if (is_null($this->profile_photo_path)) {
+            return;
+        }
+
         Storage::disk($this->profilePhotoDisk())->delete($this->profile_photo_path);
 
         $this->forceFill([
